Ajustando Exception - Consulta

+++ src/main/java/com/example/Medico/services/ConsultaService.java
@@ -1,21 +1,18 @@
 package com.example.Medico.services;
 
 import com.example.Medico.dtos.ConsultaDTO;
+import com.example.Medico.exceptions.ConsultaNotFoundException;
 import com.example.Medico.mappers.ConsultaMapper;
 import com.example.Medico.models.Consulta;
 import com.example.Medico.repository.ConsultaRepository;
 import com.example.Medico.validator.ValidaConflitoHorario;
 import com.example.Medico.validator.ValidaPaciente;
-import com.example.Medico.validator.ValidadorConsulta;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class ConsultaService {
@@ -23,7 +20,11 @@ public class ConsultaService {
     @Autowired
     private ConsultaRepository consultaRepository;
 
+    @Autowired
+    private ValidaPaciente validaPaciente;
 
+    @Autowired
+    private ValidaConflitoHorario validaConflitoHorario;
 
     public Page<ConsultaDTO> findAll(Pageable pageable) {
         return consultaRepository.findAll(pageable)
@@ -34,12 +35,6 @@ public class ConsultaService {
         return consultaRepository.findById(id)
                 .map(ConsultaMapper::convertToDTO);
     }
-
-    @Autowired
-    private ValidaPaciente validaPaciente;
-
-    @Autowired
-    private ValidaConflitoHorario validaConflitoHorario;
 
     public ConsultaDTO save(ConsultaDTO consultaDTO) {
         try {
@@ -54,9 +49,9 @@ public class ConsultaService {
     }
 
     public void deleteById(Long id) {
+        if (!consultaRepository.existsById(id)) {
+            throw new ConsultaNotFoundException("Consulta não encontrada para o id: " + id);
+        }
         consultaRepository.deleteById(id);
     }
-
-
-
 }
